FPGA_v2/requirements.txt
# Python Requirements for DPC FPGA Project
# Install with: pip install -r requirements.txt

Pillow>=9.0.0
numpy>=1.20.0
